rgbPixel initialization now takes either two tuple/lists, one tuple/list and the color (or coordinate) as separate variables, or all of the variables separately in sequence. Condition checking checks for all 4 possible states.
rgbPixel now has a setColor() function, which can take either a list/tuple of RGB values, or RGB values as separate arguments. It also has a getColor() function.
rgbPixel now has a getX() and getY() function, as X and Y are now private variables.

rgbImage now has a getAllPixels() function, which gets all of the pixels in an array.
rgbImage's reset() function has been updated for the new initialization routine for rgbPixel